requirements.txt
torch
torchvision
timm
tqdm
pyyaml
numpy
pandas
matplotlib
